requirements.txt
pygame~=2.1.2
numpy~=1.23.3
Pillow~=9.2.0